Move the gpt query to its own file
