+++ .env.example
@@ -1,2 +1,2 @@
 NEXT_PUBLIC_GOOGLE_MAPS_API_KEY=
-OPENAI_API_KEY=
+NEXT_PUBLIC_OPENAI_API_KEY=
